common c refactor
Try to resolve conflicts with chipid_cleanup branch
