+++ contributors.txt
@@ -7,8 +7,6 @@ Andrea Mucignat
 Andrew Andrianov [necromant]
 Andrey Yurovsky
 Andy Isaacson
-Andreas Sandberg [andysan]
-Antoine Faure [antoinefaure]
 Anton [Ant-ON]
 Áron Radics
 A. Sheaff
@@ -26,7 +24,6 @@ Chris Samuelson
 Christian Deussen [nullsub]
 Christophe Levantis
 Craig Lilley
-Crest [Crest]
 Dan Dev
 Dan Hepler
 Daniel Campoverde [alx741]
