+++ src/stlink-lib/usb.c.bak
@@ -1134,6 +1134,8 @@ stlink_t *stlink_open_usb(enum ugly_loglevel verbose, enum connect_type connect,
 #endif
 
     libusb_device **list = NULL;
+    // TODO: We should use ssize_t and use it as a counter if > 0.
+    // As per libusb API: ssize_t libusb_get_device_list (libusb_context *ctx, libusb_device ***list)
     ssize_t cnt = libusb_get_device_list(slu->libusb_ctx, &list);
     struct libusb_device_descriptor desc;
