+++ src/win32/unistd/unistd.h.bak
@@ -52,6 +52,7 @@
 
 #define ssize_t int
 #ifndef SSIZE_MAX
+//#define SSIZE_MAX ((sizeof(ssize_t) == 4) ? 2147483647 : 9223372036854775807)
 #define SSIZE_MAX ((sizeof(ssize_t) == 4) ? 1073741824 : ‭4611686018427387904‬)
 #endif
