+++ stlinkv1_macos_driver/install.sh
@@ -2,6 +2,9 @@
 
 ISMACOS=$(sw_vers -productVersion)
 case $ISMACOS in
+10.14*)
+    KEXT="stlink_shield_10_14.kext"
+    ;;
 10.15*)
     KEXT="stlink_shield_10_15.kext"
     ;;
