+++ stlinkv1_macos_driver/stlink_shield_10_14.kext/Contents/PkgInfo
@@ -1,0 +1,1 @@
+KEXT????
